run prettier.
