+++ .prettierignore
@@ -12,6 +12,7 @@ node_modules
 
 # Files
 login.php
+createaccount.php
 
 # Styles
 admin/bootstrap/pace/templates/pace-theme-loading-bar.tmpl.css
